requirements.txt
# Core UI
streamlit>=1.30.0
streamlit-calendar>=1.4.0
streamlit-lottie>=0.0.5

# Data & Visualization
pandas>=2.0.0
numpy>=1.24.0
altair>=5.0.0
pillow>=10.0.0

# LLM & RAG
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
crewai>=0.30.0
langchain-ollama>=0.1.0

# Embeddings & Vector Stores
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
chromadb>=0.4.0

# PDF Processing
pypdf>=3.0.0

# API
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
